secure-python-client/requirements.txt
kafka-python>=2.2.11